ensure each lambda function would have a separate name and belongs to a separate service

+++ scripts/compare_bench.sh
@@ -8,16 +8,24 @@ if [ -z $AGENT_COMMIT ]; then
     exit 1
 fi
 
+# Ensure the function has a idenifiable name
+if [ -z $NAME ]; then
+    echo "NAME not set"
+    exit 1
+fi
+
 rm -rf scripts/.cache
 rm -rf scripts/.src
 
 export AGENT_COMMIT=$AGENT_COMMIT
-export INVOKES=30
+export FUNCTION_NAME="Without-$NAME"
+export INVOKES=100
 
 ./scripts/build_publish_deploy_bench.sh
 
+export FUNCTION_NAME="With-$NAME"
 export AGENT_COMMIT=$AGENT_COMMIT^
-export INVOKES=30
+export INVOKES=100
 
 rm -rf scripts/.cache
 rm -rf scripts/.src
